- Deleted hardcoded code - Reduction rate still needs to be automated

+++ Bildverkleinerung/script.sh
@@ -1,33 +1,47 @@
 #!/bin/bash
 # Name: Bildverkleinerung
 # Datum: 17.12.2023
-# Autoren: Adrian Orlamünde, Alessandro Melcher
+# Autoren: Ruben Gonzalez Cruz, Adrian Orlamünde, Alessandro Melcher
 
 # ------------------------------- Variablen erstellen -------------------------------
 IMAGE_PATH="gbssg_logo.png"
 IMAGE_NAME="gbssg_logo"
-VERKLEINERUNGSRATE = $1
+
 
 # ------------------------------- Buckets erstellen -------------------------------
-BUCKET_NAME_SRC="bildverkleinerungsource"
-BUCKET_NAME_DST="bildverkleinerungdestination"
+BUCKET_NAME_SRC="bildverkleinerungsource1234"
+BUCKET_NAME_DST="bildverkleinerungdestination1234"
 
 REGION="us-east-1"
 
-echo "Erster Bucket wird erstellt"
-aws s3 rm s3://$BUCKET_NAME_SRC --recursive
-aws s3api delete-bucket --bucket $BUCKET_NAME_SRC --region $REGION
-aws s3api wait bucket-not-exists --bucket $BUCKET_NAME_SRC --region $REGION
-aws s3api create-bucket --bucket $BUCKET_NAME_SRC --region $REGION
-echo "Erster Bucket erstellt"
+if aws s3api head-bucket --bucket "$BUCKET_NAME_SRC" 2>/dev/null; then
+    echo "Bucket $bucket_name existiert bereits und wird zuerst gelöscht."
+    echo "Neuer, erster Bucket wird erstellt"
+    aws s3 rm s3://$BUCKET_NAME_SRC --recursive
+    aws s3api delete-bucket --bucket $BUCKET_NAME_SRC --region $REGION
+    aws s3api wait bucket-not-exists --bucket $BUCKET_NAME_SRC --region $REGION
+    aws s3api create-bucket --bucket $BUCKET_NAME_SRC --region $REGION
+    echo "Neuer, erster Bucket $BUCKET_NAME_SRC erstellt"
+else
+    echo "Erster Bucket wird erstellt"
+    aws s3api create-bucket --bucket $BUCKET_NAME_SRC --region $REGION
+    echo "Erster Bucket $BUCKET_NAME_SRC erstellt"
+fi
 
 
-echo "Zweiter Bucket wird erstellt"
-aws s3 rm s3://$BUCKET_NAME_DST --recursive
-aws s3api delete-bucket --bucket $BUCKET_NAME_DST --region $REGION
-aws s3api wait bucket-not-exists --bucket $BUCKET_NAME_DST --region $REGION
-aws s3api create-bucket --bucket $BUCKET_NAME_DST --region $REGION
-echo "Zweiter Bucket erstellt"
+if aws s3api head-bucket --bucket "$BUCKET_NAME_DST" 2>/dev/null; then
+    echo "Bucket $bucket_name existiert bereits und wird zuerst gelöscht."
+    echo "Neuer, erster Bucket wird erstellt"
+    aws s3 rm s3://$BUCKET_NAME_DST --recursive
+    aws s3api delete-bucket --bucket $BUCKET_NAME_DST --region $REGION
+    aws s3api wait bucket-not-exists --bucket $BUCKET_NAME_DST --region $REGION
+    aws s3api create-bucket --bucket $BUCKET_NAME_DST --region $REGION
+    echo "Neuer, erster Bucket $BUCKET_NAME_DST erstellt"
+else
+    echo "Erster Bucket wird erstellt"
+    aws s3api create-bucket --bucket $BUCKET_NAME_DST --region $REGION
+    echo "Erster Bucket $BUCKET_NAME_DST erstellt"
+fi
 
 
 
@@ -43,17 +57,25 @@ echo "Zweiter Bucket erstellt"
 ACCOUNTID=$(aws sts get-caller-identity --query "Account" --output text)
 LABROLE="arn:aws:iam::$ACCOUNTID:role/LabRole"
 
-LAMBDA_FUNCTION_NAME="LamdaBildkomprimierung"
+LAMBDA_FUNCTION_NAME="LamdaBildkomprimierung1234"
 
-echo "Lambda-Funktion wird erstellt"
-aws lambda delete-function --function-name "$LAMBDA_FUNCTION_NAME"
-echo "Lambda-Funktion wird gelöscht"
-aws lambda create-function --function-name "$LAMBDA_FUNCTION_NAME" --zip-file fileb://lambda_function.zip --handler lambda_function.lambda_handler --runtime python3.11 --role $LABROLE
-echo "Lambda-Funktion erstellt"
+
+if aws lambda get-function --function-name "$LAMBDA_FUNCTION_NAME" 2>/dev/null; then
+    echo "Lambda-Function $LAMBDA_FUNCTION_NAME existiert bereits und wird zuerst gelöscht."
+    aws lambda delete-function --function-name "$LAMBDA_FUNCTION_NAME"
+    echo "Lambda-Funktion wird erstellt"
+    aws lambda create-function --function-name "$LAMBDA_FUNCTION_NAME" --zip-file fileb://pythonfunction.zip --handler lambda_function.lambda_handler --runtime python3.8 --role $LABROLE
+    echo "Lambda-Funktion erstellt"
+else
+    echo "Lambda-Funktion wird erstellt"
+    aws lambda create-function --function-name "$LAMBDA_FUNCTION_NAME" --zip-file fileb://pythonfunction.zip --handler lambda_function.lambda_handler --runtime python3.8 --role $LABROLE
+    echo "Lambda-Funktion erstellt"
+fi
+
+LAMBDA_ARN=$(aws lambda get-function-configuration --function-name "$LAMBDA_FUNCTION_NAME" --query "FunctionArn" --output text --region "$REGION")
+
 
 echo "Trigger wird erstellt"
-LAMBDA_FUNCTION_ARN=$(aws lambda get-function --function-name "$LAMBDA_FUNCTION_NAME" --query "Configuration.FunctionArn" --output text)
-
 aws lambda add-permission \
     --function-name $LAMBDA_FUNCTION_NAME \
     --action lambda:InvokeFunction \
@@ -61,12 +83,12 @@ aws lambda add-permission \
     --source-arn arn:aws:s3:::$BUCKET_NAME_SRC \
     --statement-id s3invoke
 
-aws s3api put-bucket-notification-configuration \
+    aws s3api put-bucket-notification-configuration \
     --bucket $BUCKET_NAME_SRC \
     --notification-configuration '{
         "LambdaFunctionConfigurations": [
             {
-                "LambdaFunctionArn": "'$LAMBDA_FUNCTION_ARN'",
+                "LambdaFunctionArn": "'$LAMBDA_ARN'",
                 "Events": ["s3:ObjectCreated:*"]
             }
         ]
@@ -77,35 +99,23 @@ echo "Trigger erstellt"
 
 
 
+
+
+
+
+
+
 # ------------------------------- Bild in den Source Bucket hochladen -------------------------------
+# Speicherpfad des Bildes
+current_directory="$(pwd)"
+
+# Durchsuche das aktuelle Verzeichnis nach einem Bild
+
+
 echo -n "Bildverkleinerungsrate eingeben: "
-read reductionrate
+reductionrate="2"
 
+echo "Bild wird in $BUCKET_NAME_SRC hochgeladen"
 # Bild in den Source Bucket hochladen
-echo "Bild wird hochgeladen"
-aws s3 cp $IMAGE_PATH s3://$BUCKET_NAME_SRC/ --region $REGION --metadata reduction_rate=$reductionrate --metadata output_bucket="s3://$BUCKET_NAME_DST/$IMAGE_NAME""_resized.png"
-echo "Bild hochgeladen"
-
-
-
-
-
-
-
-# ------------------------------- Neuens Bild lokal abspeichern -------------------------------
-# Setze hier deine AWS-Region und S3-Bucket-Informationen ein
-OBJEKT_SCHLUESSEL="$IMAGE_NAME"".png"
-
-# Setze hier den lokalen Zielordner ein
-LOKALER_ORDNER="$(pwd)/Bildverkleinert"
-
-# Bild von S3 nach lokal kopieren
-aws s3 cp s3://$BUCKET_NAME_DST/$OBJEKT_SCHLUESSEL $LOKALER_ORDNER --region $REGION
-
-# Ausgabe, ob der Kopiervorgang erfolgreich war
-if [ $? -eq 0 ]; then
-    echo "Bild erfolgreich von S3 nach $LOKALER_ORDNER kopiert."
-else
-    echo "Fehler beim Kopieren des Bildes von S3."
-fi
-
+aws s3 cp $IMAGE_PATH s3://$BUCKET_NAME_SRC/ --region $REGION --metadata reduction_rate=2 --metadata output_bucket=$BUCKET_NAME_DST
+echo "Bild in $BUCKET_NAME_SRC hochgeladen"
